+++ packages/pros/src/lib.rs
@@ -393,6 +393,7 @@ pub mod prelude {
             screen::{Circle, Line, Rect, Screen, Text, TextFormat, TextPosition, TouchState},
             smart::{
                 distance::DistanceSensor,
+                expander::AdiExpander,
                 gps::GpsSensor,
                 imu::InertialSensor,
                 link::{Link, RxLink, TxLink},