+++ CHANGELOG.md
@@ -82,6 +82,7 @@ Before releasing:
 - All ADI device bindings (#55)
 - `LocalKey` now has `Cell`/`RefCell`-specific methods for setting and taking values. (#42)
 - `Peripherals` and `DynamicPeripherals` structs to ensure that you have only registered one device on a given smart or ADI port. (#53)
+- Support for ADI Expander modules with `AdiExpander`. (#63)
 
 ### Fixed